migrate notification center state management to zustand
